+++ ondam-backend/src/main/java/com/piveguyz/ondambackend/diaryRecord/command/application/controller/DiaryRecordController.java
@@ -17,7 +17,7 @@ public class DiaryRecordController {
     }
 
     @PostMapping("/sendDiary")
-    public ResponseEntity<String> sendDiary(@RequestParam("id") Integer id){
+    public ResponseEntity<String> sendDiary(@RequestParam("id") Long id){
         boolean result = diaryRecordService.sendDiary(id);
         if(result) {
             return ResponseEntity.status(HttpStatus.ACCEPTED).body("일기가 전송되었습니다.");
